sem_6/Database_course_project/Database_course_backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework
markdown
django-filter
Pillow